Use bitset rather than hashset for iterative clique finding
One logic is, only the vertices that can form a k-clique have a chance
to form a (k+1)-clique. For all the neighbors of a specific vertex a,
while enumerating its k-clique, we record those vertices that can form
(k+1)-clique with it using a bitset.

I used to use hashset, which is not good because it is not
memory-efficient, and furthermore, the hashset is not sorted, so every
time I have to sort the survived vertices, which is (to some extent)
costly and unnecessary.